move depinject specific module items to depinject
