+++ depinject/appconfig/config.go
@@ -13,8 +13,8 @@ import (
 	"sigs.k8s.io/yaml"
 
 	appv1alpha1 "cosmossdk.io/api/cosmos/app/v1alpha1"
-	"cosmossdk.io/core/internal"
 	"cosmossdk.io/depinject"
+	internal "cosmossdk.io/depinject/internal/appconfig"
 )
 
 // LoadJSON loads an app config in JSON format.
