+++ depinject/internal/appconfig/buf.yaml
