+++ types/module/core_module.go
@@ -15,6 +15,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
@@ -194,7 +195,7 @@ func (c coreAppModuleBasicAdaptor) RegisterLegacyAminoCodec(amino *codec.LegacyA
 
 // RegisterServices implements HasServices
 func (c coreAppModuleBasicAdaptor) RegisterServices(cfg Configurator) {
-	if module, ok := c.module.(appmodule.HasServices); ok {
+	if module, ok := c.module.(HasGRPCServices); ok {
 		err := module.RegisterServices(cfg)
 		if err != nil {
 			panic(err)
