+++ types/module/module.go
@@ -38,8 +38,10 @@ import (
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
 	"github.com/spf13/cobra"
 	"golang.org/x/exp/maps"
+	"google.golang.org/grpc"
 
 	"cosmossdk.io/core/appmodule"
+	am "cosmossdk.io/core/appmodule"
 	"cosmossdk.io/core/genesis"
 	errorsmod "cosmossdk.io/errors"
 	storetypes "cosmossdk.io/store/types"
@@ -214,6 +216,26 @@ type HasInvariants interface {
 type HasServices interface {
 	// RegisterServices allows a module to register services.
 	RegisterServices(Configurator)
+}
+
+// HasServices is the extension interface that modules should implement to register
+// implementations of services defined in .proto files.
+type HasGRPCServices interface {
+	am.AppModule
+
+	// RegisterServices registers the module's services with the app's service
+	// registrar.
+	//
+	// Two types of services are currently supported:
+	// - read-only gRPC query services, which are the default.
+	// - transaction message services, which must have the protobuf service
+	//   option "cosmos.msg.v1.service" (defined in "cosmos/msg/v1/service.proto")
+	//   set to true.
+	//
+	// The service registrar will figure out which type of service you are
+	// implementing based on the presence (or absence) of protobuf options. You
+	// do not need to specify this in golang code.
+	RegisterServices(grpc.ServiceRegistrar) error
 }
 
 // HasConsensusVersion is the interface for declaring a module consensus version.
@@ -458,13 +480,6 @@ func (m *Manager) RegisterServices(cfg Configurator) error {
 	for _, module := range m.Modules {
 		if module, ok := module.(HasServices); ok {
 			module.RegisterServices(cfg)
-		}
-
-		if module, ok := module.(appmodule.HasServices); ok {
-			err := module.RegisterServices(cfg)
-			if err != nil {
-				return err
-			}
 		}
 
 		if cfg.Error() != nil {
