+++ types/module/module_test.go
@@ -654,7 +654,7 @@ func (MockCoreAppModule) ExportGenesis(ctx context.Context, target appmodule.Gen
 }
 
 var (
-	_ appmodule.AppModule   = MockCoreAppModule{}
-	_ appmodule.HasGenesis  = MockCoreAppModule{}
-	_ appmodule.HasServices = MockCoreAppModule{}
+	_ appmodule.AppModule    = MockCoreAppModule{}
+	_ appmodule.HasGenesis   = MockCoreAppModule{}
+	_ module.HasGRPCServices = MockCoreAppModule{}
 )
