+++ x/counter/module.go
@@ -21,8 +21,7 @@ import (
 var (
 	_ module.AppModuleBasic = AppModule{}
 
-	_ appmodule.AppModule   = AppModule{}
-	_ appmodule.HasServices = AppModule{}
+	_ appmodule.AppModule = AppModule{}
 )
 
 // AppModuleBasic defines the basic application module used by the consensus module.
